Restructure plugin folder for easy zip and release
