+++ QNSPECT/processing/algorithms/__init__.py
