+++ QNSPECT/processing/algorithms/compare_scenarios/comparison_utils.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import sys
 
-sys.path.append(Path(__file__).parent.parent)
+sys.path.append(Path(__file__).parents[1])
 from qnspect_utils import perform_raster_math
