+++ QNSPECT/processing/algorithms/load_run/load_run.py
@@ -28,7 +28,7 @@ from qgis.core import (
 import processing
 import json
 
-from QNSPECT.qnspect_algorithm import QNSPECTAlgorithm
+from QNSPECT.processing.qnspect_algorithm import QNSPECTAlgorithm
 
 
 class LoadPreviousRun(QNSPECTAlgorithm):
