+++ QNSPECT/processing/algorithms/modify_land_cover/modify_land_cover_by_field.py
@@ -31,7 +31,7 @@ from qgis.core import (
 )
 import processing
 
-from QNSPECT.qnspect_algorithm import QNSPECTAlgorithm
+from QNSPECT.processing.qnspect_algorithm import QNSPECTAlgorithm
 
 
 class ModifyLandCover(QNSPECTAlgorithm):
