+++ QNSPECT/processing/algorithms/rasterize_soil/rasterize_soil.py
@@ -31,7 +31,7 @@ from qgis.core import (
 )
 import processing
 
-from QNSPECT.qnspect_algorithm import QNSPECTAlgorithm
+from QNSPECT.processing.qnspect_algorithm import QNSPECTAlgorithm
 
 
 class RasterizeSoil(QNSPECTAlgorithm):
