+++ QNSPECT/processing/algorithms/run_analysis/relief_length_ratio.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from qgis.core import QgsUnitTypes, QgsRasterLayer, QgsProcessing
 import processing
 
-sys.path.append(str(Path(__file__).parent.parent))
+sys.path.append(str(Path(__file__).parents[1]))
 from qnspect_utils import perform_raster_math
 
 __all__ = ("create_relief_length_ratio_raster",)
