+++ QNSPECT/resources/style-colors.json
